Refactor Order models to use BigDecimal for storing price values instead of Int

+++ core/domain/src/main/java/ru/livetyping/zarina/core/domain/model/order/Order.kt
@@ -1,5 +1,6 @@
 package ru.livetyping.zarina.core.domain.model.order
 
+import java.math.BigDecimal
 import java.time.LocalDate
 
 // Marked as stable on config/compose/stability_config.txt
@@ -9,7 +10,7 @@ public sealed class Order {
     public abstract val productCount: Int
     public abstract val date: LocalDate
     public abstract val status: OrderStatus
-    public abstract val totalPrice: Int
+    public abstract val totalPrice: BigDecimal
 
     public val isPaid: Boolean get() = status == OrderStatus.PAID
