+++ core/domain/src/main/java/ru/livetyping/zarina/core/domain/model/order/OrderPrice.kt
@@ -1,8 +1,10 @@
 package ru.livetyping.zarina.core.domain.model.order
 
+import java.math.BigDecimal
+
 // Marked as stable on config/compose/stability_config.txt
 public data class OrderPrice(
-    val orderPrice: Int,
-    val deliveryPrice: Int,
-    val totalPrice: Int,
+    val orderPrice: BigDecimal,
+    val deliveryPrice: BigDecimal,
+    val totalPrice: BigDecimal,
 )
