+++ core/domain/src/main/java/ru/livetyping/zarina/core/domain/model/order/OrderShort.kt
@@ -1,6 +1,7 @@
 package ru.livetyping.zarina.core.domain.model.order
 
 import ru.livetyping.zarina.core.domain.model.common.Url
+import java.math.BigDecimal
 import java.time.LocalDate
 
 // Marked as stable on config/compose/stability_config.txt
@@ -10,7 +11,7 @@ public data class OrderShort(
     override val productCount: Int,
     override val date: LocalDate,
     override val status: OrderStatus,
-    override val totalPrice: Int,
+    override val totalPrice: BigDecimal,
     val products: List<Product>,
 ) : Order() {
     // Marked as stable on config/compose/stability_config.txt
